requirements.txt
mido
python-dotenv
